+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-dev:
+1.26.3:
   - provide support for additional S3 store options
   - show error when attempting to delete non-filesystem wallets
+  - provide additional support for Capella
 
 1.26.2
   - remove check that requires capella prior to generating validator credentials change operations